Annotation search: fix annotator and time constraint filter logic
The annotator and time constraint filters have to be valid for all results.
This, however, wasn't the case so far. Both were used as alternative filters to
the annotations, while they should actually constrain the annotation even more.
This commit fixes this by making sure annotation, annotator and time filters are
combined with an AND to query the search results.