Tracing: store position of active node in unscaled stack coordinates
Storing the active tree node's position in scaled stack space
coordinates causes a problem when the zoom-level is changed: If one
doesn't re-select the active node its position isn't updated, which
would be required, because it is zoom-level dependent. To fix this, it
seems more cleaner to me to generally store the coordinates of the
active tree node in unscaled stack space rather than observing a stack's
scale change.

This change fixes #515, a 3D viewer issue which has been there for quite
a while: Centering the active node after changing the zoom-level would
result in a wrong position. That was because the 3D viewer expected the
active tree node's position to be in scaled stack space coordinates, but
these were not updated when the zoom-level was adjusted through the
slider.