Review system: add missing review updates on split and join
When skeletons are joined or split, threre is always one skeleton removed or
created, respectively. In these situations all reviews of the treenodes
belonging to the removed or created treenode need to be updated, because the
skeleton ID is part of the review table. The column is there for performance
reasons and has to be kept up-to-date when a tree node moves to a different
skelton.

This fixes #670. Thanks to Loni Brueckmann for reporting this problem.