3D viewer: don't save reviews as node properties
Attaching the review information to each node hits performance significantly.
Therefore, this commit stores the returned review information as property of
each skeleton object. Shading methods can access this field, if they need it.
This also simplifies the delayed loading of the review information, which will
be added soon.

This is related to issue #668.