Add migration to ensure consistensy of review table
Due to a bug (that was fixed recently) it could be the case that there
are reviews in the review table that link to a wrong skeleton. The
problem was that a review's skeleton reference was not updated when
skeletons were merged or split. With the help of the migration added in
this commit, the review table is made consistent. That is, all reviews
are checked and if the skeleton they reference is another one that their
treenode references, the review is updated to reference the treenode's
skeleton.