Review widget: fix respecting followed users on next segment selection
If a checkbox for a user on top of a column is checked, their reviews will be
respected when the next segment is selected. This means if they have reviewed a
segment by a 100 percent, it will be skipped for selecting the next segment for
review. Unfortunately, this did not really work before and this commit fixes it.

Thanks to Javier Valdes Aleman for reporting this bug.