Make post-synaptic nodes display confidence values
There was a problem that post-synaptic confidence values were not
displayed along the edges to the connector. The reason was that the
confidence value wasn't read from the given parameter, but was expected
to be a field of the arrow object. This wasn't correct and so this
commit changes this to use the confidence parameter. The post-synaptic
confidence values are now shown again.

This is part of the fix for issue #662. Thanks to Davi Bock for
reporting this problem.