Neuron navigator: support annotation removal from multiple neurons
This commit adds a new button to the annotation page of of the neuron
navigator. Sitting on top of the table of neurons that are annotated
with the current annotation, this button allows one to remove the
annotation from the all selected neurons. A user is required to have
permission to remove the annotation from *all* neurons, before any
annotation is actually removed.

If there are co-annotations selected, a button is shown for every
annotation contributing to the current filter set.

This is related to issue #654.