Annotation removal: don't require permissions for all links in request
When annotations are un-linked from a list of entities, one was required
to have permissions to remove the annotation from every entity, before any
actual removal.  This commit changes this behavior: Now all links are
removed for which a user has permissions, the others remain untouched.

This was discussed in issue #654.