Graph widget: invalidate Cytoscape's canvas dimensions on resize
Telling Cytoscape explicitly about the resize is required, because it
doesn't monitor the dimensions of the viewport (which makes sense,
because it can be expensive). This fixes an issue which, since the last
Cytoscape update, caused the canvas to stay at the same size as before,
when the widget size was changed.