Graph widget: reduce the redraw timeout after resize to 100ms
This will cause a redraw to happen faster after a resize, but keeps
resizing still smooth. The former value of 400ms felt a bit to large
if one clicked right away into the graph after a resize. If Cytoscape
wasn't informed about the dimension change yet, the mouse positions were
wrong. This doesn't happen if the update is scheduled faster.