Selection table: hide color wheel properly for new rows
The DIV containing the color wheel should be hidden by default---not
only the color wheel itself. This wasn't the case before and it caused a
line break right in front of the info button. With this commit the DIV
is hidden when added and no more line break is shown.

Thanks to Javier Valdes Aleman for pointing me to this little annoyance.