utils.js: define fetchCompactSkeletons
Fetches each skeleton given an array of IDs, one at a time to avoid time outs
and blocking a server thread for too long.
Dispatches returned JSON and failures and done events via handler functions.

It's a refactoring of a function previously in the webglapp.js,
for use in other widgets.