Connectivity widget: fix check-box behavior
As reported in issue #688, the selection check-boxes didn't keep their
state when the threshold was changed. This commit fixes this by
maintaining the selection state of each skeleton (and check-box) in as
member field of the widget itself. Thus, when refreshed, the state isn't
lost and can be reused. This fixes #688.