Fix general focus catching of UI (e.g. un-focusing inputs)
The UI object has a catchFocus() method to un-focus the currently active
element (e.g. a text input field). This is done by focusing an empty
anchor element (the 'focusCatcher'). This element doesn't accept any
input and so all keyboard and mouse events can bubble up. Calling
focus() on this element does, however, only work if it is a proper
anchor, i.e. if it has a 'href' attribute. This commit adds this
'href' attribute and (un-)focusing now works properly.

Thanks to Nils Holger Kranenburg for reporting this issue. The original
problem leading to this was the following: When jumping to a specific
slice by entering its slice number into the z-index input box of the
Navigator/Move tool followed by pressing Enter, the focus would remain
in this box. So even after one clicked into the image to e.g.  pan, one
couldn't use the plus and minus keys to change the scale. After this
commit, the input box is un-focused after clicking the mouse on the
image.