Group graph widget: fix adjacency matrix export
Since commit 7674945, where the master branch was merged into the
group-graph branch, the export of the adjacency matrix from within the
group graph widget wasn't working anymore. In the branch, the name
generation for the exported neurons was changed and two bugs were
introduced: 1. map was used instead of reduce, but the parameters were
not adjusted and 2. The closing quote for the neuron name got lost. Both
are fixed by this commit and the adjacency matrix export should work
again.

Thanks to Philipp Schlegel for reporting this issue.