Graph widget: fix adjacency matrix export
The adjacency matrix export was broken since the introduction of typed arrays.
Unfortunately, typed arrays only support a very limited set of array operations
right now (e.g. no reduce, map, join, etc.). A join, however, was used for the
adjacency matrix export---this commit replaces this join with a for-loop.