Arbor.js: new function resampleSlabs
Leaves all root/branch/end nodes in place, and resamples the nodes
in the slabs to a desired node interdistance.

The last edge of each slab (prior to reaching a branch node or an end node)
will be equal or shorter than the desired node interdistance.

The resampling is based on a Gaussian convolution.