WebGLApp: introducing options to resample the skeleton.
Root, branch and end nodes are not altered. Slabs are resampled so that
nodes have a maximum, constant interdistance of "resampling_delta", in nm.

Reduces e.g. a neuron with 1000 nodes to about 50 nodes when resampling
to a delta of 5000 nm.

At extremely large deltas, the arbor is reduced to root, branch and end
nodes, linked by straight lines between them.