Add /tests URL to give access to simple QUnit front-end tests
This commit adds a Django template, a URL to call it and an example file
to have a simple way to create tests for the client-side JavaScript of
CATMAID. QUnit is used for this. Tests should be placed in files within
CATMAID's 'static/tests/' directory. New files need to be included in
the 'templates/catmaid/tests.html' file.