SynapseClustering: speed up, and fix an error.
The error was at:
 if (delta_density[id] < 0) return;
... which now, being a loop instead of a forEach, the return
should have been a continue.

The speed up is based on using Arbor.allNeighbors() rather than allSuccessors()
which avoids having to generate the neighbors with an Array.clone operation
for every treenode.

Additionally partitions are now iterated with a for loop rather than a forEach,
which removes levels of closure lookup.