SynapseClustering: new function findAxonCut
SynapseClustering is the best namespace in which to put this function
that is used by both the CircuitGraphPlot and the WebGLApplication
to compute, given an array of nodes with maximum centrifugal flow centrality,
the node at which to optimally cut.

Includes inspecting nodes for presence of output synapses, rather
than merely selecting the highest-order end node of the subarbors defined
within the array of max centrality nodes.