Reorder columns of location based tables
This commit re-creates the table layout of the location, treenode, connector and
region_of_interest tables. The tables are optimized for data alignment to waste
as little space as possible. As discussed in issue #723 this saves 66 Bye per
row.

After this migration is applied (which could take some minutes, depending on the
size of the location based tables), you should run VACUUM FULL from  with
Postgres.