Add migration to remove duplicate annotations
To do all duplicate annotations are retrieved and looked at. If one duplicate
annotation was created by the same person who created the annotated neuron, this
annotation is kept, the other(s) are marked for removal. If such a user doesn't
exist, the oldest annotation link is used.

The administrator is presented a list of all annotations that will be removed
along with the neuron ID and the user who created that annotation link. One can
then confirm the removal, ignore these duplicates or cancel the migration.

Issue #724 discusses this issue.