User stats: use Agg backend for matplotlib
This backend can be used when we only want to have images in memory or
save them to files. This is the case for CATMAID's user statistics.
Changing this allows us to run the user statistics code in the
development server. This wasn't possible before due to threading issues.
These originated in the default matplotlib backend, which is based on
Tk/TkInter. For more details on matplotlib backends see:
http://matplotlib.org/faq/usage_faq.html#what-is-a-backend