Cropping: add maximum file size setting
This commit adds a maximum allowed file size to the settings and the cropping
module. It can be adjusted by changing the GENERATED_FILES_MAXIMUM_SIZE setting
(in Bytes) which defaults to 50MB. The cropping module is estimating the final
file size by summing up the size of each image part it needs to create the final
image stack. The size of a single part is estimated by multiplying the size of
the tile it is referring to with the ratio of actually used area of it. This
fixes #730.