Don't require Shift key to rename active neuron with F2
When the Object Tree was still around, F2 was used to rename nodes in
it. When the Shift key was pressed additionally, the active neuron was
renamed. We don't use the Object Tree anymore and this commit therefore
changes this behavior: F2 is now used to rename the active neuron
(without even looking at the Shift key).