SVG overlay: only process clicks targeted at own view
If one opens another stack of the same project while in tracing mode, it was
possible to create new nodes by clicking in the non-active (!) stack. This could
cause unwanted nodes, possibly even outside of the extent of the active stack,
making the new nodes unreachable. This commit fixes this by making sure a click
event is only processed in an overlay (which triggers the actual node
creation) if its view is the very same the click was targeted at. The event is
not stopped from bubbling up at the moment to make it possible to handle it at
other places. Currently, this triggers the activation of another stack/widget if
there is any.

This closes #419.