Connectivity widget: use DataTables for partner tables
Extend both upstream and downstream partner tables with DataTables in
the connectivity widget. This allows for client-side sorting, row
filtering, and column reordering. Because the partner tables use complex
headers, a fork of the ColReorder DataTables plugin was created:

http://github.com/aschampion/ColReorder

in branch

clean/support-tree-complex-headers

Commit 2f6d1b386ca165f40216199c58b3f62a3d8927fd is used for the
dependencies in this commit, which is based on ColReorder 1.1.2.

Element IDs for the upstream and downstream partner tables are changed
to be unique, for proper DataTables operation. Table CSS is modified to
support DataTables' dynamic styling. The table collapse functionality
now only triggers when the caret is clicked, to allow clicks in the
header cell for sorting.

Closes acardona/CATMAID#694 (column reordering).
Closes acardona/CATMAID#736 (column sorting and filtering).
Closes acardona/CATMAID#620 (stacked partner table spacing).