Action: prevent recursive image loading error
Action buttons use the onError event to fallback to a PNG icon if an SVG
icon is not found. However, if a PNG icon is not found either, a
recursive error chain is created, crippling the browser. Instead, null
out the error handler after the first call, allowing the fallback source
to be set but preventing onError recursion.