Console: prevent status bar from overlapping coordinates
Long status bar messages were shown behind the coordinate display in the
status bar, making both illegible. Instead, long status bar messages are
now truncated with an ellipsis. Hovering the mouse over the console will
remove the coordinate display and show the entire status bar, expanding
the console over the rest of the UI if necessary.

To make the width of the status bar adjust to the width of the
coordinate display, the coordinate display is reordered to be the first
child of the console. Effective console height is reduced from 35 to 29
px. The slice controls toggle is the same z-order as the context map
toggle so that the console can occlude it when hovered.

Tested on current OSX Chrome, Firefox and Safari.

Closes acardona/CATMAID#739.