Remove CSS files for Internet Explorer 5, 6 and 7
Currently we don't support Internet Explorer at all and some features and
libraries we use are not supported in IE 5, 6 or 7. For instance, Array's map,
reduce, filter, etc. functions are introduced in IE 9. Libraries like Cytoscape
and parts of CATMAID use the canvas element. Support for it was introduced in IE
9 as well. The 3D viewer uses WebGL, which was introduced in IE 11. Therefore,
there is certainly no need for CSS files that improve support for IE 5, 6 and 7.