Connectivity widget: ability to export partner tables
Adds the DataTables TableTools extension to allow export of connectivity
partner tables to CSV or PDF files or as plaintext to the clipboard.
Note that TableTools uses the summary (last) header row as the only
header in exports, so a fork of TableTools 2.2.2 was created at:

http://github.com/aschampion/TableTools.git

on branch

support-multirow-headers

at commit 2b2cc10dda12a1831610805c105007d799f70eef using the patch from:

http://goo.gl/AWvuOb

and included in this commit. This fork includes the first two header
rows (i.e., the headers except for the UI summary row) in all exports.