Make neuron name serivce a singleton
Up to now the neuron name service was instantiated only once and stored in a
global variable. This, however, isn't very robust. This commit makes the neuron
name service a singleton which can be retrieved using

    NeuronNameService.getInstance()

Getting rid of the global variable also allows easier minification of this code,
in case we want to use this.