Export widget: add tree geometry export
Add a trivial JSON export of treenode and connector geometry, as
requested for quick export and hacking. This is a munge of the backend
compact-skeleton API for accessibility. Improvements to the export
format and implementation are welcome.

Cherry-pick of 96c1c4b3d817ef20df0308b512245d731f0afb24 by request.