Review widget: queue selection of next segment after pending requests
The selection of the next segment depends on client side information about the
review state of all nodes of all segments. To select the next segment based on
the latest information, the selection is now queued after currently pending
requests. This makes sure all single node review requests have been handled
before.

This is related to issue #743. Though it isn't clear yet if this was the only
problem that could lead to this behavior.