Review widget: move comment and remove unneeded user name override
By default, the anonymous user (i.e. when no-one is logged in) has ID -1. Back
when the review system didn't support multiple users, -1 as a user ID was also
used to mark a treenode as not reviewed. This has changed and is not done
anymore, because there is now a separate review table containing only actual
reviews. Therefore, we don't need this name override for non-reviewed treenode
entries anymore. Besides this, the name also had a typo in it.