Rename VNC-Browser's treelines.js file due to name collision
Unfortunately, there is no way to have application specific rules in
django-pipeline. All patterns are therefore tested against the static files of
all applications. This can lead to name collisions if a file exists in two
applications. This is what happened with treeline.js and it is therefore renamed
in the vncbrowser application.