Allow admin to add a range of broken slices in one go
When a new broken slice is added from within the Django admin interface,
a new optional field is now displayed: last index. If used, all broken
slice entries in the range [index, last index] will be created
autmatically if they don't exist already. This makes is possible to add
a range of broken slices at once, like it was discussed in #714.