Tracing tool: inherit basic actions from Navigator
... instead of defining them all over again.

This commit removes the navigation actions (zoom, pan) from the
Tracing tool and instead gets them from Navigator.

It also removes a keycode ambiguity between '.' and 'DEL'. Moving
down in z now only works with the actual '.' key and not with the one
on the numeric keypad. Before, in Navigator, one could also move down
in z with the 'DEL/.' key on the numeric keypad. However, this key is
used to delete the active node in the Tracing tool.