Neuron Dendrogram: add tool infrastructure and skeleton loading
The neuron dendrogram will eventually display a neuron as a tree structure (i.e.
a dendrogram). This commit adds basic tool infrastructure to create its window
and to open it by pressing F4 while in the tracing tool. A button will follow.

Also, the functionality to retrieve the nodes of a skeleton is added and to
transform it into a tree structure that is known to D3.