Remove duplicate and conflicting definition of ResizeHandle
There were two definitions of ResizeHandle, one in a separate file and
one in windowmanager.js. Before django-pipeline was used, the latter one
was used. However, after django-pipeline was included it was not
deterministic anymore which definition took precedence. That is because
all files get concatenated into one. Apparently the compiler choose
sometimes (usually) the wrong one. This commit fixes this by moving the
definition from within windowmanager.js into the separate file,
effectively overriding the second definition and removing the first one.