Move treenode table button to neuron navigator
The treenode table isn't used very often and does not need the prominent
place it has right now. It is therefore moved to the neuron node of the
neuron navigator. With the help of a button one can now open it from there for
the neuron shown in the node.