Compartment graph: fix cytoscape SVG export
We recently found out that the SVG export was broken, but couldn't really find a
commit that caused this. Even commits that were known to work in the past
suddenly failed. As it turns out (after stepping for quite a while through
Cytoscape and SVGKit) does Cytoscape use Path2D to cache paths if it is
available. SVGKit, however, is not able to use it and silently fails to draw
paths if Cytoscape uses Path2D. This in turn makes our export function crash.

The reason why this didn't show up earlier is that Chrome only recently marked
Path2D to be not experimental anymore and made it available in the global
namespace. Firefox, had it enabled for quite some time already, but we
apparently never tested the SVG export in it. It now works there as well.

This is related to the discussion in cytoscape/cytoscape.js#639.