Notification table: fix initialization of action column
The Datatables API changed and we apparently didn't see this bug before:
Initializing the action column didn't work anymore, because fnRender
isn't supported anymore and the replacement doesn't provide the row
index anymore. This commit fixes this problem and makes the notification
table load again.

Thanks to Comfrey McCarthy and Willie Tobin for reporting this problem.