Neuron deletion: fix broken deletion of neurons with reviews
If a neuron is deleted, its reviews should be removed as well.
Unfortunately, we can't rely on cascading deletes from the data
base and so we have to delete the reviews manually within our
deletion transaction. This manual deletion is added by this commit.

This change Relates to issue #763. Thanks to Comfrey McCarthy and
Willie Tobin for pointing me at this.