Add small and simple event bus
This event system can be used to decouple different parts in CATMAID.
For instance, being notified about the change of the active skeleton or
active node is easily possible with the Events object. The producer (who
actually changes the active node) would trigger the event:

  Event.trigger('active-node-changed');

A Consumer, like maybe the 3D widget, that is interested in such an
event would register a callback for this event:

  Event.on('active-node-changed', callback);

A QUnit test has been added.