Tests: fix transactional tests
All TransactionTests after the first run one were failing because of a
duplicate user_id key error when loading the UserProfile fixture. This
was because the Guardian __init__ was loading an AnonymousUser between
each test, which triggered the post_save signal on User, creating a
UserProfile with id 2 for user_id -1. This post_save occurs without the
'raw' flag (unlike for loaddata from the fixtures), so it cannot be
prevented.  Since the id is different from the UserProfile fixture and
the save method for UserProfile which finds and updates UserProfiles
attempting to duplicate user_ids is not called for fixture loading, the
duplicate key IntegrityError occurred.

Since we are not preventing User fixtures from firing the post_save
signal, just remove the UserProfile from the fixtures and the
create_user_profile post_save hook do its job.

Remove many TransactionTest tests that were not testing any CATMAID
functionality. Left test_fail_unexpectedly as a sanity test, though it
is still of dubious value.