SVG Overlay: switch from Raphael to D3
Switch from Raphael to D3 for the SVG overlay in order to work directly
with the SVG elements to improve performance and enable additional
functionality.

This migration will be iterative over the next several commits. Also,
this is a direct port to D3, without code restructuring, so it is not
idiomatic D3 code.