Neuron dendrogram: subscribe to changes of active node and highlight it
If users now select a treenode in the tracing layer, the neuron
dendrogram will select this node as well as its sub-tree. This also
changes how double clicking a node in the dendrogram works: It now only
changes the active node and the highlighting is done as response to this
change.