Reviews: support review of sub-arbors
If the new "Start to review current sub-arbor" button is pressed, only
the nodes downstream along with the selected node will be reviewed. This
commit implements this on the back-end to have it available right away.
Eventually, this will move into the front-end.

Issue #796 is basically implemented with this commit, but will be left
open until the front-end implementation is in place.