Add alternative mode of radius editing
When the 'o' key is pressed, a new mode of entering the radius for a
node is activated: A circle is drawn around the current node of which
one can modify its radius with the movement of the mouse. If 'o' is
pressed again, the regular radius editing dialog is opened, with the
radius field prepopulated with the radius of circle drawn before. The
conversion to nanometers takes into account that there might be
different resolutions in the X and Y components of the point where
users had their mouse last. To make this effect more clear we might want
to replace the circle with a line. This would require the user to
explicitly state which components play a role in the conversion to
nanometers (by explicitly pointing to a position).

If Shift is pressed along with 'o', the dialog is shown as before,
without any drawing.