Statistics: increase connector count only for own completed connectors
So far the connector count in the statistics widget was incremented for
every connector node that was created by a user in a given time frame.
This is changed to an incremented count when connector relations (link
between pre and post) are finishing by a user. This should motivate
finishing synapses early.