SynapseClustering.findAxonCut: improved heuristic
Currently will look for the lowest-order node on the 50% distal
half of the set of nodes in the high plateau of e.g. centrifugal
synapse flow centrality, where the node either has an output synapse
or is a branch node where more than one of the children branches
has output synapses, or the most distal node if none of the above.