GroupGraph: growing respects the min_downstream and min_upstream.
I've renamed variables from min_pre to min_downstream
and min_post to min_upstream for clarity. The real issue
was that "all" and "none" were expressed as -1 and 0, respectively,
and that was not taken in to account in the client side.
Additionally min_pre implied downstream but meant upstream
and viceversa, so that is reversed now. The new wording helps.

Additionally, paths are found by double-breadth first search,
therefore min_downstream and min_upstream msut both be defined
and set to the maximum of both, for consistency.