Fix missing permission check on delete of treenodes
The treenode deletion method so far only checked if a user has
permissions to edit the neuron the treenode is part of. However, a uer
could still delete the treenode if the neuron wasn't locked or if a user
was in the group of the user who locked the neuron, even if the user
didn't own the treenode.

This commit changes this behavior so that a user who wants to delete a
treenode must have permission to edit the whole neuron (as before) and
the user must have permission to edit the single treenode (i.e. created
it or be part of the group of the creator).