SVG overlay: fix catch radius
Because the stroke of an SVG circle is centered along its radius, the
desired effective catch radius must be doubled. While this replicates
the previous effective catch radius of ~8, note that current values for
treenode cannot be made larger because of a Chrome bug:

https://code.google.com/p/chromium/issues/detail?id=239860

Thanks to Kathi Eichler for the bug report.