Fix small problem with skeleton source select controls
It sporadically happened that the select control couldn't be
initialized. This usually happened to the 3D viewer and the problem
seems to be that the update code for these selects assumes, that there
is a selection already. This, however, is not always the case. Using the
first element as a fall-back fixes this problem for me.