Neuron navigator: don't explicitly refresh node after annotation changes
There is no need for explicitly refreshing the current node if an annotation
has been added or removed. The neuron navigator is registered with the neuron
name service, which will notify its clients after it has been updated. Adding or
removing an annotation triggers a refresh of the name service.

Nevertheless, it would be nice to have some kind of semaphore pattern in use
that stops the navigator from breaking if there are multiple select_node()
calls after another. The problem is Datatables that can't handle a returning
AJAX request if the actual table was removed already (due to the second
select_node() call).

Fixes #829