3D viewer: add sorting by different labeling for SVG catalog export
This allows to select an alternative labeling for sorting the skeletons in the
SVG catalog export. For labeling individual results, the general neuron name
service is still in use. However, for sorting (and therefore determining the
cell in the output table) one can now choose a different labeling scheme.
Another neuron name service instance is used for creating the names and
therefore the options are same. In contrast to the regular name service, no
fallback list is used and if a labeling cannot be applied to a neuron (e.g. to
to the lack of annotations with a chosen meta annotation), an error is raised.