Update jQuery to v1.8.3
This was required due to to a bug in the version we used before. It came into
effect when exporting a neuron catalog SVG and raised an error.

I went through the list of removed and changed functions [1] and found only
smaller things. Since these are rather minor issues, I'll commit them
separately.

[1] http://blog.jquery.com/2012/08/09/jquery-1-8-released/