Remove unused jQuery.dd widget
This "msDropDown" UI element was added in a40a70f621 and is not used
anymore (the back then created CATMAID widget also doesn't exist
anymore).

Besides not being used, the widget in its current version is also not
compatible with jQuery 1.8.3, to which we just moved. The deprecated

  .data("events")

was removed and this was used by jquery.dd.