3D viewer: reduce coordinate precision in SVG export
The precision used for coordinates in the exported SVG was overkill. For the
regular SVG export, one decimal digit is now used. This is also the default for
the catalog export. However, there it can be adjusted in the presented dialog.

Currently, only the precision of lines is reduced, because there is no other
object we use at the moment.

Relates to issue acardona/CATMAID#840.